WIP : Physics
https://trello.com/c/kC96eLRV
Use 64-bit integer UUID for culling data.

+++ Source/Engine/PhysicsSystem/IPhysicsSystem.h
@@ -19,7 +19,7 @@ struct CullingData
 	VisiblilityType visiblilityType;
 	MeshUsageType meshUsageType;
 	CullingDataChannel cullingDataChannel;
-	uint32_t UUID;
+	uint64_t UUID;
 };
 
 struct BVHNode
